Refactoring code; separating functionality
