+++ src/microhttpd/Makefile.am
@@ -139,7 +139,10 @@ endif
 
 if ENABLE_HTTP2
 libmicrohttpd_la_SOURCES += \
-  http2/h2_connection.c http2/h2_upgrade.c http2/h2.h
+  http2/h2_config.c http2/h2_upgrade.c \
+  http2/h2_connection.c \
+  http2/h2_session.c http2/h2_stream.c \
+  http2/h2_helper.c http2/h2.h
 endif
