+++ src/microhttpd/connection_https.c
@@ -211,7 +211,7 @@ MHD_run_tls_handshake_ (struct MHD_Connection *connection)
                  ALPN_HTTP_2_0_LENGTH)) )
       {
         selected.data = (unsigned char *)ALPN_HTTP_2_0;
-        MHD_set_h2_callbacks (connection);
+        h2_set_h2_callbacks (connection);
       }
     else
 #endif /* HTTP2_SUPPORT */
@@ -219,7 +219,7 @@ MHD_run_tls_handshake_ (struct MHD_Connection *connection)
         /* Default HTTP version */
         selected.data = (unsigned char *)ALPN_HTTP_1_1;
 #ifdef HTTP2_SUPPORT
-        MHD_set_h1_callbacks (connection);
+        h2_set_h1_callbacks (connection);
 #endif /* HTTP2_SUPPORT */
       }
